+++ include/Matrix.h
@@ -11,7 +11,7 @@ namespace COAL
 
         [[nodiscard]] constexpr Matrix4(){};
 
-        [[nodiscard]] constexpr Matrix4(const double *array, _maybe_unused int array_size)
+        [[nodiscard]] constexpr Matrix4(const float *array, _maybe_unused int array_size)
         {
             assert(16 == array_size);
 
@@ -20,22 +20,22 @@ namespace COAL
                     this->_matrix[i][j] = array[i * 4 + j];
         }
 
-        [[nodiscard]] constexpr Matrix4(const std::array<std::array<double, 4>, 4> &array)
+        [[nodiscard]] constexpr Matrix4(const std::array<std::array<float, 4>, 4> &array)
         {
             for (char i = 0; i < 4; i++)
                 for (char j = 0; j < 4; j++)
                     this->_matrix[i][j] = array[i][j];
         }
 
-        [[nodiscard]] constexpr Matrix4(const std::vector<std::vector<double>> &values)
+        [[nodiscard]] constexpr Matrix4(const std::vector<std::vector<float>> &values)
         {
             for (char i = 0; i < 4; i++)
                 for (char j = 0; j < 4; j++)
                     this->_matrix[i][j] = values[i][j];
         }
 
-        // constructor for (double, double, double, int, double, double, double, int, double, double, double, double, double, double, double, double)
-        [[nodiscard]] constexpr Matrix4(const double a00, const double a01, const double a02, const double a03, const double a10, const double a11, const double a12, const double a13, const double a20, const double a21, const double a22, const double a23, const double a30, const double a31, const double a32, const double a33)
+        // constructor for (float, float, float, int, float, float, float, int, float, float, float, float, float, float, float, float)
+        [[nodiscard]] constexpr Matrix4(const float a00, const float a01, const float a02, const float a03, const float a10, const float a11, const float a12, const float a13, const float a20, const float a21, const float a22, const float a23, const float a30, const float a31, const float a32, const float a33)
         {
             this->_matrix[0][0] = a00;
             this->_matrix[0][1] = a01;
@@ -56,19 +56,19 @@ namespace COAL
         }
 
         // // [][] operator
-        // [[nodiscard]] constexpr double &operator[](const int index) noexcept
+        // [[nodiscard]] constexpr float &operator[](const int index) noexcept
         // {
         //     return this->_matrix[index / 4][index % 4];
         // }
 
         // [] operator
-        [[nodiscard]] constexpr double *operator[](const int index) noexcept
+        [[nodiscard]] constexpr float *operator[](const int index) noexcept
         {
             return this->_matrix[index];
         }
 
         // / operator
-        [[nodiscard]] constexpr Matrix4 operator/(const double rhs) const noexcept
+        [[nodiscard]] constexpr Matrix4 operator/(const float rhs) const noexcept
         {
             return Matrix4(
                 this->_matrix[0][0] / rhs, this->_matrix[0][1] / rhs, this->_matrix[0][2] / rhs, this->_matrix[0][3] / rhs,
@@ -118,8 +118,8 @@ namespace COAL
             return result;
         }
 
-        // * by std::vector<std::vector<double>>
-        [[nodiscard]] constexpr Matrix4 operator*(const std::vector<std::vector<double>> &other) const noexcept
+        // * by std::vector<std::vector<float>>
+        [[nodiscard]] constexpr Matrix4 operator*(const std::vector<std::vector<float>> &other) const noexcept
         {
             Matrix4 result;
 
@@ -145,7 +145,7 @@ namespace COAL
         }
 
         // * operator
-        [[nodiscard]] constexpr Matrix4 operator*(const double &other) const noexcept
+        [[nodiscard]] constexpr Matrix4 operator*(const float &other) const noexcept
         {
             Matrix4 result;
 
@@ -194,7 +194,7 @@ namespace COAL
         }
 
         // *= operator
-        [[nodiscard]] constexpr Matrix4 &operator*=(const double &other) noexcept
+        [[nodiscard]] constexpr Matrix4 &operator*=(const float &other) noexcept
         {
             *this = *this * other;
 
@@ -218,13 +218,13 @@ namespace COAL
         }
 
         // [] operator
-        [[nodiscard]] constexpr const double &operator[](const int index) const noexcept
+        [[nodiscard]] constexpr const float &operator[](const int index) const noexcept
         {
             return this->_matrix[index / 4][index % 4];
         }
 
         // [] operator
-        [[nodiscard]] constexpr const double &operator()(const int row, const int column) const noexcept
+        [[nodiscard]] constexpr const float &operator()(const int row, const int column) const noexcept
         {
             return this->_matrix[row][column];
         }
@@ -242,7 +242,7 @@ namespace COAL
         }
 
         // determinant
-        [[nodiscard]] constexpr double determinant4() const noexcept
+        [[nodiscard]] constexpr float determinant4() const noexcept
         {
             return this->_matrix[0][0] * this->_matrix[1][1] * this->_matrix[2][2] * this->_matrix[3][3] +
                    this->_matrix[0][0] * this->_matrix[1][2] * this->_matrix[2][3] * this->_matrix[3][1] +
@@ -271,7 +271,7 @@ namespace COAL
         }
 
         template <size_t size>
-        _nodiscard constexpr void sub_matrix(const double (&matrix)[size + 1][size + 1], const int row, const int column, double (&sub_matrix)[size][size]) const
+        _nodiscard constexpr void sub_matrix(const float (&matrix)[size + 1][size + 1], const int row, const int column, float (&sub_matrix)[size][size]) const
         {
             bool a_replaced = false;
             bool b_replaced = false;
@@ -292,26 +292,26 @@ namespace COAL
             }
         }
 
-        [[nodiscard]] double constexpr determinant(const double (&matrix)[2][2]) const
+        [[nodiscard]] float constexpr determinant(const float (&matrix)[2][2]) const
         {
             return matrix[0][0] * matrix[1][1] - matrix[0][1] * matrix[1][0];
         }
 
-        [[nodiscard]] double constexpr determinant(const double (&matrix)[3][3]) const
+        [[nodiscard]] float constexpr determinant(const float (&matrix)[3][3]) const
         {
-            double det = 0;
+            float det = 0;
 
             for (char i = 0; i < 3; i++)
             {
                 if (i % 2 == 0)
                 {
-                    double temp_matrix[2][2];
+                    float temp_matrix[2][2];
                     sub_matrix<2>(matrix, i, 0, temp_matrix);
                     det += matrix[i][0] * determinant(temp_matrix);
                 }
                 else
                 {
-                    double temp_matrix[2][2];
+                    float temp_matrix[2][2];
                     sub_matrix<2>(matrix, i, 0, temp_matrix);
                     det -= matrix[i][0] * determinant(temp_matrix);
                 }
@@ -329,10 +329,10 @@ namespace COAL
             {
                 for (int j = 0; j < 4; j++)
                 {
-                    double temp_matrix[3][3];
+                    float temp_matrix[3][3];
                     sub_matrix<3>(_matrix, i, j, temp_matrix);
 
-                    cofactor[i][j] = pow(-1, i + j) * determinant(temp_matrix);
+                    cofactor[i][j] = (float)pow(-1, i + j) * determinant(temp_matrix);
                 }
             }
 
@@ -351,9 +351,9 @@ namespace COAL
             return this->adjugate() / this->determinant4();
         }
 
-        _nodiscard Matrix4 translate(std::vector<double> const &values) const
+        _nodiscard Matrix4 translate(std::vector<float> const &values) const
         {
-            std::vector<std::vector<double>> translation_values = {
+            std::vector<std::vector<float>> translation_values = {
                 {1, 0, 0, values[0]},
                 {0, 1, 0, values[1]},
                 {0, 0, 1, values[2]},
@@ -363,9 +363,9 @@ namespace COAL
             return *this * translation_values;
         }
 
-        _nodiscard Matrix4 translate(const double x, const double y, const double z) const
+        _nodiscard Matrix4 translate(const float x, const float y, const float z) const
         {
-            std::vector<std::vector<double>> translation_values = {
+            std::vector<std::vector<float>> translation_values = {
                 {1, 0, 0, x},
                 {0, 1, 0, y},
                 {0, 0, 1, z},
@@ -375,9 +375,9 @@ namespace COAL
             return *this * translation_values;
         }
 
-        _nodiscard Matrix4 scale(std::vector<double> const &values) const
+        _nodiscard Matrix4 scale(std::vector<float> const &values) const
         {
-            std::vector<std::vector<double>> translation_values = {
+            std::vector<std::vector<float>> translation_values = {
                 {values[0], 0, 0, 0},
                 {0, values[1], 0, 0},
                 {0, 0, values[2], 0},
@@ -387,9 +387,9 @@ namespace COAL
             return *this * translation_values;
         }
 
-        _nodiscard Matrix4 scale(const double x, const double y, const double z) const
+        _nodiscard Matrix4 scale(const float x, const float y, const float z) const
         {
-            std::vector<std::vector<double>> translation_values = {
+            std::vector<std::vector<float>> translation_values = {
                 {x, 0, 0, 0},
                 {0, y, 0, 0},
                 {0, 0, z, 0},
@@ -399,9 +399,9 @@ namespace COAL
             return *this * translation_values;
         }
 
-        _nodiscard Matrix4 rotate_x(const double radians) const
+        _nodiscard Matrix4 rotate_x(const float radians) const
         {
-            std::vector<std::vector<double>> rotate_x_values = {
+            std::vector<std::vector<float>> rotate_x_values = {
                 {1, 0, 0, 0},
                 {0, std::cos(radians), -std::sin(std::cos(radians)), 0},
                 {0, std::sin(std::cos(radians)), std::cos(radians), 0},
@@ -411,9 +411,9 @@ namespace COAL
             return *this * rotate_x_values;
         }
 
-        _nodiscard Matrix4 rotate_y(const double radians) const
+        _nodiscard Matrix4 rotate_y(const float radians) const
         {
-            std::vector<std::vector<double>> rotate_y_values = {
+            std::vector<std::vector<float>> rotate_y_values = {
                 {std::cos(radians), 0, std::sin(radians), 0},
                 {0, 1, 0, 0},
                 {-std::sin(radians), 0, std::cos(radians), 0},
@@ -423,9 +423,9 @@ namespace COAL
             return *this * rotate_y_values;
         }
 
-        _nodiscard Matrix4 rotate_z(const double radians) const
+        _nodiscard Matrix4 rotate_z(const float radians) const
         {
-            std::vector<std::vector<double>> rotate_z_values = {
+            std::vector<std::vector<float>> rotate_z_values = {
                 {std::cos(radians), -std::sin(radians), 0, 0},
                 {std::sin(radians), std::cos(radians), 0, 0},
                 {0, 0, 1, 0},
@@ -435,14 +435,14 @@ namespace COAL
             return *this * rotate_z_values;
         }
 
-        _nodiscard Matrix4 rotate(const double radians_x, const double radians_y, const double radians_z) const
+        _nodiscard Matrix4 rotate(const float radians_x, const float radians_y, const float radians_z) const
         {
             return this->rotate_x(radians_x).rotate_y(radians_y).rotate_z(radians_z);
         }
 
-        _nodiscard Matrix4 shear(double Xy, double Xz, double Yx, double Yz, double Zx, double Zy) const
+        _nodiscard Matrix4 shear(float Xy, float Xz, float Yx, float Yz, float Zx, float Zy) const
         {
-            std::vector<std::vector<double>> shear_values = {
+            std::vector<std::vector<float>> shear_values = {
                 {1, Xy, Xz, 0},
                 {Yx, 1, Yz, 0},
                 {Zx, Zy, 1, 0},
@@ -478,7 +478,7 @@ namespace COAL
         }
 
     private:
-        double _matrix[4][4] = {{1, 0, 0, 0}, {0, 1, 0, 0}, {0, 0, 1, 0}, {0, 0, 0, 1}};
+        float _matrix[4][4] = {{1, 0, 0, 0}, {0, 1, 0, 0}, {0, 0, 1, 0}, {0, 0, 0, 1}};
     };
 
     constexpr const Matrix4 IDENTITY = Matrix4();
