+++ include/Patterns/Gradient.h
@@ -16,21 +16,21 @@ namespace COAL
 
         [[nodiscard]] COAL::Color color_at(const COAL::Point &p) const override
         {
-            // double x = p.x;
-            // double y = p.y;
-            // double z = p.z;
+            // float x = p.x;
+            // float y = p.y;
+            // float z = p.z;
 
-            // double r = (x * m_transform[0][0] + y * m_transform[0][1] + z * m_transform[0][2] + m_transform[0][3]) /
+            // float r = (x * m_transform[0][0] + y * m_transform[0][1] + z * m_transform[0][2] + m_transform[0][3]) /
             //            (x * m_transform[3][0] + y * m_transform[3][1] + z * m_transform[3][2] + m_transform[3][3]);
-            // double g = (x * m_transform[1][0] + y * m_transform[1][1] + z * m_transform[1][2] + m_transform[1][3]) /
+            // float g = (x * m_transform[1][0] + y * m_transform[1][1] + z * m_transform[1][2] + m_transform[1][3]) /
             //            (x * m_transform[3][0] + y * m_transform[3][1] + z * m_transform[3][2] + m_transform[3][3]);
-            // double b = (x * m_transform[2][0] + y * m_transform[2][1] + z * m_transform[2][2] + m_transform[2][3]) /
+            // float b = (x * m_transform[2][0] + y * m_transform[2][1] + z * m_transform[2][2] + m_transform[2][3]) /
             //            (x * m_transform[3][0] + y * m_transform[3][1] + z * m_transform[3][2] + m_transform[3][3]);
 
             // return COAL::Color(r, g, b);
 
             Color distance = m_second_color - m_first_color;
-            double fraction = p.x - std::floor(p.x);
+            float fraction = p.x - std::floor(p.x);
 
             return m_first_color + distance * fraction;
         }
